fix(build): stop client bundle pulling child_process via versionCheck
The Kimi sponsor banner is a client component whose gate logic
(kimiSponsorBanner.ts) imported isNewer/normalizeVersion from
lib/system/versionCheck.ts. That module also imports node:child_process
for the npm-CLI version lookup, so Next.js tried to bundle child_process
for the browser and failed:

    Module not found: Can't resolve 'child_process'
    ./src/lib/system/versionCheck.ts (19:1)

Extract the two pure, Node-free semver helpers into a new
lib/system/versionCompare.ts and re-export them from versionCheck.ts so
server-side importers are unchanged. Point the client-side
kimiSponsorBanner.ts at the pure module, severing the child_process path
from the browser bundle. No behavior change.
